add module 4

+++ Week3/Lesson15/spring-security-jwt/src/main/java/com/codegym/controller/UserController.java
@@ -57,7 +57,7 @@ public class UserController {
 
     /* ---------------- DELETE USER ------------------------ */
     @RequestMapping(value = "/users/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<String> deleteUserById(@PathVariable int id) {
+    public ResponseEntity<String> deleteUserById(@PathVariable Long id) {
         userService.delete(id);
         return new ResponseEntity<>("Deleted!", HttpStatus.OK);
     }
